# docs/swagger.yaml
definitions:
  main.User:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /signin:
    get:
      consumes:
      - application/json
      description: Authenticate a user with a username and password
      parameters:
      - description: User credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.User'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error parsing JSON
          schema:
            type: string
        "401":
          description: Password is incorrect
          schema:
            type: string
        "404":
          description: No user with that username
          schema:
            type: string
        "405":
          description: Invalid request method
          schema:
            type: string
        "500":
          description: Failed to get a response from database" or "Error generating
            token
          schema:
            type: string
      summary: Sign in an existing user
  /signup:
    post:
      consumes:
      - application/json
      description: Create a new user with a username and password
      parameters:
      - description: User credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.User'
      produces:
      - application/json
      responses:
        "200":
          description: User with username {username} was created
          schema:
            type: string
        "400":
          description: Error parsing JSON
          schema:
            type: string
        "409":
          description: An user with the provided username already exists
          schema:
            type: string
        "500":
          description: Error hashing password" or "Failed to get a response from database
          schema:
            type: string
      summary: Sign up a new user
swagger: "2.0"